🧹 refactor tar connection
