--- providers/os/connection/tar/connection.go
@@ -1,26 +1,0 @@
-// Copyright (c) Mondoo, Inc.
-// SPDX-License-Identifier: BUSL-1.1
-
-package tar
-
-/*
-func (c *TarConnection) Identifier() (string, error) {
-	return c.PlatformIdentifier, nil
-}
-
-func (c *TarConnection) Labels() map[string]string {
-	return c.Metadata.Labels
-}
-
-func (c *TarConnection) PlatformName() string {
-	return c.Metadata.Name
-}
-
-*/
-
-// FIXME: How to handle this now?
-// func (c *TarConnection) PlatformIdDetectors() []providers.PlatformIdDetector {
-// 	return []providers.PlatformIdDetector{
-// 		providers.TransportPlatformIdentifierDetector,
-// 	}
-// }
